requirements.txt
# Day 08 Lab — RAG Pipeline
# Cài đặt: pip install -r requirements.txt

# === Core ===
python-dotenv>=1.0.0

# === LLM Providers ===
openai>=1.0.0                    # Option A: OpenAI API
google-generativeai>=0.5.0       # Option B: Google Gemini
anthropic>=0.88.0                # Option C: Anthropic Claude (nhóm dùng — claude-sonnet-4-5)

# === Vector Store ===
chromadb>=0.5.0                  # Vector database (local, không cần server)

# === Embeddings ===
voyageai>=0.3.7                  # Voyage AI embeddings (nhóm dùng — voyage-multilingual-2)
# sentence-transformers>=2.7.0   # Fallback local nếu không có API key

# === Sprint 3: Sparse/Hybrid Retrieval ===
rank-bm25>=0.2.2                 # BM25 keyword search

# === Sprint 3: Rerank ===
# sentence-transformers đã bao gồm CrossEncoder

# === Utils ===
tqdm>=4.66.0                     # Progress bar khi index nhiều docs

# === PDF -> Markdown (CLI + Web) ===
Flask>=3.0.0
requests>=2.31.0
tiktoken>=0.9.0
unstructured[pdf]>=0.16.19
unstructured-inference>=0.8.0
httpx<0.28.0

# === PDF -> Markdown (CLI + Web) ===
Flask>=3.0.0
requests>=2.31.0
tiktoken>=0.9.0
unstructured[pdf]>=0.16.19
unstructured-inference>=0.8.0
httpx<0.28.0


#Để chạy giao diện demo
gradio>=4.0.0